Added ability to dynamically build State Machine with the builder pattern
Added SetSuperState to primary State interface to allow for setting the super state when defining the state through the builder pattern

Modified scope of protected methods in AutomatonymousStateMachine to `protected internal` to allow the internal builders to access the methods without having to re-create them

Added new methods for defining State and Event to AutomatnymousStateMachine required by the builder pattern

Added Modify method, BuilderStateMachine class, and static Build method to AutomatonymousStateMachine to allow for the builder pattern to be initiated through the primary state machine. A proposed approach for a simple way to make it dynamic.

Added new interfaces and internal classes for the builder pattern.

Copied 165 of the 170 current tests, categorized them as "Dynamic Modify" and altered all state machine instances to utilize the builder pattern instead to ensure specs are still maintained.